fix tests running

+++ Jenkinsfile
@@ -4,29 +4,44 @@ node {
       cleanWs()
       git branch: 'development', url: 'https://github.com/luomus/laji.git'
       sh 'npm ci'
+      sh 'mkdir test-results'
+      sh 'rm -rf test-results/*'
     }
     stage('Quality') {
-      sh 'npm run lint'
+      catchError(buildResult: 'FAILURE', stageResult: 'FAILURE') {
+        sh './node_modules/.bin/ng lint laji --format junit > ./test-results/output-tslint.xml'
+      }
     }
-    stage('Run tests') {
-      sh 'node ./node_modules/protractor/bin/webdriver-manager update'
-      sh 'npm run e2e'
-      junit 'projects/laji/e2e/test-results/**/*.xml'
+    stage('Run integration tests') {
+      sh './node_modules/.bin/webdriver-manager clean'
+      sh './node_modules/.bin/webdriver-manager update --versions.chrome=$(google-chrome --version | awk \'{print $3}\')'
+      catchError(buildResult: 'FAILURE', stageResult: 'FAILURE') {
+        sh 'npm run e2e:ci -- --webdriver-update=false'
+      }
+      junit allowEmptyResults: true, testResults: '**/test-results/**/*.xml'
     }
     stage('Build') {
-      milestone()
-      sh 'rm -rf dist'
-      sh 'npm run --silent build:dev'
-      sh 'pre-compress-web-assets dist/browser'
+      if (currentBuild.result == 'SUCCESS') {
+        milestone()
+        sh 'rm -rf dist'
+        sh 'npm run --silent build:dev'
+        sh 'pre-compress-web-assets dist/browser'
+      }
     }
     stage('Archive') {
-      sh 'tar -cvzf dist.tar.gz --strip-components=1 dist'
-      archive 'dist.tar.gz'
+      if (currentBuild.result == 'SUCCESS') {
+        sh 'tar -cvzf dist.tar.gz --strip-components=1 dist'
+        archive 'dist.tar.gz'
+      }
     }
     stage('Deploy staging') {
-      milestone()
-      sh 'scp -r dist node@192.168.10.26:/data/dev_laji_fi/'
-      slackSend(color: 'good', message: "dev.laji.fi - #$BUILD_NUMBER Success")
+      if (currentBuild.result == 'SUCCESS') {
+        milestone()
+        sh 'scp -r dist node@192.168.10.26:/data/dev_laji_fi/'
+        slackSend(color: 'good', message: "dev.laji.fi - #$BUILD_NUMBER Success")
+      } else {
+        slackSend(color: 'danger', message: "dev.laji.fi - #$BUILD_NUMBER Failed")
+      }
     }
   }
 }
